fix(scrape): restore downloading album by slug
The feature to download by slug was broken in the rewrite, any attempt
to use it would crash with a memory exception when attempting to read a
member of a nil http.Response.

The DownloadPage function violates the invariant that an
http.Client.Do-like function would return one of http.Response or error
by returning nil for both if the server returned non-200 status codes.

This changeset modifies RetrieveAlbum to construct the proper album URL,
and handles an error response from DownloadPage before attempting to
defer closing the body. DownloadPage itself is modified to return an
error if an unexpected status is received.